add route to landing
