+++ src/components/HeaderPreview.js
@@ -1,15 +1,16 @@
 import React from 'react';
-import logo from './../images/logo_awesome.svg'
+import { Route, Switch, Link } from 'react-router-dom';
+
+import logo from './../images/logo_awesome.svg';
 
 const HeaderPreview = () => {
-    return (
-        <header class="header">
-            <a href="index.html">
-                <img class="header__image" src={logo} alt="Crea tu tarjeta personal" />
-            </a>
-        </header>
-    )
+  return (
+    <header class='header'>
+      <Link to='/'>
+        <img class='header__image' src={logo} alt='Crea tu tarjeta personal' />
+      </Link>
+    </header>
+  );
 };
 
 export default HeaderPreview;
-
